+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.6.0 - 2025-04-12
+
 ### Changed
 
  - Put all the integer type shortcuts into the new `Int_ty` module.